+++ src/app/layout.tsx
@@ -7,7 +7,7 @@ import NavigationBar from '@/components/NavigationBar'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
+  title: 'DnF | 내 캐릭터 찾기, 랭킹보기',
   description: 'Generated by create next app',
 }
 
@@ -21,7 +21,7 @@ export default function RootLayout({
       <body className={inter.className}>
         <StoreProvider>
           <NavigationBar />
-          <main className="pt-32">{children}</main>
+          <main className=" pt-28">{children}</main>
         </StoreProvider>
       </body>
     </html>
