+++ src/service/utils/BuffJobCategory.ts
@@ -1,0 +1,6 @@
+export const BuffJobCategory = [
+  { job: '크루세이더', jobId: 'f6a4ad30555b99b499c07835f87ce522' },
+  { job: '크루세이더(여)', jobId: '0c1b401bb09241570d364420b3ba3fd7' },
+  { job: '인첸트리스', jobId: '3909d0b188e9c95311399f776e331da5' },
+  { job: '뮤즈', jobId: 'b9cb48777665de22c006fabaf9a560b3' },
+]
