+++ src/service/utils/JobCategory.ts
@@ -3,6 +3,7 @@ export const category = [
     job: '귀검사(남)',
     jobId: '41f1cdc2ff58bb5fdc287be0db2a8df3',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '웨펀마스터', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '소울브링어', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '버서커', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -14,6 +15,7 @@ export const category = [
     job: '귀검사(여)',
     jobId: '1645c45aabb008c98406b3a16447040d',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '소드마스터', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '다크템플러', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '데몬슬레이어', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -25,6 +27,7 @@ export const category = [
     job: '거너(남)',
     jobId: 'afdf3b989339de478e85b614d274d1ef',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '레인저', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '런처', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '메카닉', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -36,6 +39,7 @@ export const category = [
     job: '거너(여)',
     jobId: '944b9aab492c15a8474f96947ceeb9e4',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '레인저', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '런처', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '메카닉', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -46,6 +50,7 @@ export const category = [
     job: '격투가(여)',
     jobId: 'a7a059ebe9e6054c0644b40ef316d6e9',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '넨마스터', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '스트라이커', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '스트리트파이터', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -56,6 +61,7 @@ export const category = [
     job: '격투가(남)',
     jobId: 'ca0f0e0e9e1d55b5f9955b03d9dd213c',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '넨마스터', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '스트라이커', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '스트리트파이터', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -66,17 +72,18 @@ export const category = [
     job: '마법사(여)',
     jobId: '3909d0b188e9c95311399f776e331da5',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '엘리멘탈마스터', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '소환사', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '배틀메이지', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '마도학자', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
-      { job: '인첸트리스', jobId: '92da05ec93fb43406e193ffb9a2a629b' },
     ],
   },
   {
     job: '마법사(남)',
     jobId: 'a5ccbaf5538981c6ef99b236c0a60b73',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '엘리멘탈바머', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '빙결사', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '블러드메이지', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -88,6 +95,7 @@ export const category = [
     job: '프리스트(남)',
     jobId: 'f6a4ad30555b99b499c07835f87ce522',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '크루세이더', jobId: '37495b941da3b1661bc900e68ef3b2c6' },
       { job: '인파이터', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '퇴마사', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
@@ -98,6 +106,7 @@ export const category = [
     job: '프리스트(여)',
     jobId: '0c1b401bb09241570d364420b3ba3fd7',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '이단심판관', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '무녀', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '미스트리스', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
@@ -107,6 +116,7 @@ export const category = [
     job: '도적',
     jobId: 'ddc49e9ad1ff72a00b53c6cff5b1e920',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '사령술사', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '쿠노이치', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '섀도우댄서', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
@@ -117,6 +127,7 @@ export const category = [
     job: '나이트',
     jobId: '0ee8fa5dc525c1a1f23fc6911e921e4a',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '카오스', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '팔라딘', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '드래곤나이트', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
@@ -127,6 +138,7 @@ export const category = [
     job: '마창사',
     jobId: '3deb7be5f01953ac8b1ecaa1e25e0420',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '듀얼리스트', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '드래고니안랜서', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '다크랜서', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
@@ -137,6 +149,7 @@ export const category = [
     job: '총검사',
     jobId: '986c2b3d72ee0e4a0b7fcfbe786d4e02',
     growJob: [
+      { job: '전체', jobId: '' },
       { job: '요원', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
       { job: '트러블슈터', jobId: '6d459bc74ba73ee4fe5cdc4655400193' },
       { job: '스폐셜리스트', jobId: 'c9b492038ee3ca8d27d7004cf58d59f3' },
@@ -146,7 +159,10 @@ export const category = [
   {
     job: '아처',
     jobId: 'b9cb48777665de22c006fabaf9a560b3',
-    growJob: [{ job: '트래블러', jobId: '618326026de1a1f1cfba5dbd0b8396e7' }],
+    growJob: [
+      { job: '전체', jobId: '' },
+      { job: '트래블러', jobId: '618326026de1a1f1cfba5dbd0b8396e7' },
+    ],
   },
   {
     job: '다크나이트',
